Refactor to return new table when indexing event
Before this change, when you indexed the event table with a key that was
not an method name, it would essentially return itself but with the
"_pattern" field set to the key used to event table. I think this
would've worked for the most part, but it would not stop you from
infinitely recursively indexing the same table:

    local t = auto.event.user.one.two.three.four.five.whatever
    {
	_event = "user",
	_pattern = "whatever",
	...
    }

This new implementation limits the indexing on the `luauto.event` module
to two or three levels:

    auto.event.<event_name>[.<pattern>]:<method>([args])

Indexing beyond will just return nil. This also allows us in the future
to limit the methods we can call on `<event_name>.<pattern>`, for
instance, a "buffer" method will throw an error because you cannot
specify the buffer when specifying the pattern and vice versa.