Add a couple methods
Add a method "get" which gets a list of autocommands matching the event
and pattern if specified.

Add a method "patterns" which returns itself but with the "_pattern"
field set to the argument it was called with. This should allow you to
set a list of patterns. I'm probably gonna have to refactor this cause
I'm thinking of adding an intermediary table access between the event
name and the pattern like: `auto.event.user.pat.example` so we'll see
how that goes.

Remove the local "api" table which was just used an alias to `vim.api`.
Using those aliases broke the mock table in the tests:(