Major changes to `luauto.group`
Create another table called "GroupProxy". Use this as the metatable for
the tables that we return when we index this submodule. This makes our
"Group" table a bit more private, make certain fields read-only, and
allows us to access the autogroup ids and names as properties in the
table instead of method calls.

i know i'm making these massive commits but i'm still trying stuff out
bear with me plz thank