Remove use of "GroupProxy" table
We can move the metamethods over to the "Group" table and add a new
"proxy" method to it. I thought it would be tricky or confusing keeping
track of `self` that's why I created a separate table to specify the
metamethods.

Also, instead of using a local table "mem" to cache the created proxy
tables, we can add them to the module itself so we don't have to make
multiple calls to the `__index` metamethod.

Add some comments and make changes to the spec file to reflect the
changes in method signatures.