Remove "EventProxy" table
Instead, move `__index`, `__newindex`, and `__call` metamethods to the
Event table and add a method "_proxy" that will return a proxy table
with the metatable set to the Event table itself.

Add logic for indexing values with case-insensitive keys.

Move the memoizing logic out to the module's `__index` metamethod.

Some changes to the spec file to reflect changes in method signatures.