Add a lot of tests
I think I'm slowly getting better at writing tests... like, I'm kinda
feeling it ngl. One thing I know was smelly about the previous tests was
that the testing blocks were way too nested and confusing.

Teachable moment (for myself mainly): keep those blocks focused on a
piece of your code and don't go too deep nesting them. If they're
getting confusing and hard to reason through them, that's a smell and we
gotta take a step back from the testing and see if we can structure it
better. Or, we gotta rethink the code itself, or whatever implementation
that we are brainstorming in our head.