Refactor and break up into separate files
Make a function "get_scoped_events" which returns a table that is used
to manage autocmds for events, in particular to get, clear, exec, or
create them. Can be scoped to an augroup and/or buffer.

We can use this function when creating the "luauto.cmd" table to manage
autocmds that are not scoped to an augorup and/or buffer.

We can also use this function when we are creating a table to manage
autocmds for events that are only in this augroup.