Bring all the submodules back together
- events
- augroup
- autocmd

The "events" submodule can create dictionaries that are used to manage
autocmds or buflocal-autocmds based on event in a particular group or
the default if non specified.

The "autocmd" submodule can manage autocmds based on event in the
default group.

The "augroup" submodule can manage autocmds in a specific augroup.