Change the "Augroup" table's `__index` field
Instead of using a function for the `__index` metamethod which supported
getting the group id from a field "id" instead of calling a method, I
think that bit of sugar is unnecessary.

To get the id of a group, we can call its "create" method and get its
return value, which reflects the behavior of calling
`nvim_create_augroup()`.